requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.20.0
Pillow>=8.0.0
tqdm>=4.50.0
pyyaml>=5.4.0
wandb>=0.12.0
matplotlib>=3.4.0
scipy>=1.6.0
pycocotools>=2.0.2